+++ libopensimcreator/Documents/MeshWarper/UndoableTPSDocumentActions.h
@@ -9,6 +9,7 @@
 #include <liboscar/Maths/Vec3.h>
 #include <liboscar/Utils/UID.h>
 
+#include <memory>
 #include <string_view>
 #include <unordered_set>
 
@@ -63,13 +64,13 @@ namespace osc
     void ActionLoadMesh(UndoableTPSDocument&, const Mesh&, TPSDocumentInputIdentifier);
 
     // prompts the user to browse for an input mesh and assigns it to the document
-    void ActionLoadMeshFile(UndoableTPSDocument&, TPSDocumentInputIdentifier);
+    void ActionLoadMeshFile(std::shared_ptr<UndoableTPSDocument>, TPSDocumentInputIdentifier);
 
     // loads landmarks from a CSV file into source/destination slot of the document
-    void ActionLoadLandmarksFromCSV(UndoableTPSDocument&, TPSDocumentInputIdentifier);
+    void ActionLoadLandmarksFromCSV(std::shared_ptr<UndoableTPSDocument>, TPSDocumentInputIdentifier);
 
     // loads non-participating landmarks from a CSV file into the source input
-    void ActionLoadNonParticipatingLandmarksFromCSV(UndoableTPSDocument&);
+    void ActionLoadNonParticipatingLandmarksFromCSV(std::shared_ptr<UndoableTPSDocument>);
 
     // saves all source/destination landmarks to a CSV file (matches loading)
     void ActionSaveLandmarksToCSV(const TPSDocument&, TPSDocumentInputIdentifier, lm::LandmarkCSVFlags = lm::LandmarkCSVFlags::None);