+++ libopensimcreator/UI/MeshWarper/MeshWarpingTabInputMeshPanel.h
@@ -481,7 +481,7 @@ namespace osc
             {
                 if (ui::draw_menu_item("Mesh File"))
                 {
-                    ActionLoadMeshFile(m_State->updUndoable(), m_DocumentIdentifier);
+                    ActionLoadMeshFile(m_State->getUndoableSharedPtr(), m_DocumentIdentifier);
                 }
                 if (ui::begin_menu("Generated Mesh")) {
                     drawGeneratedMeshOptions(SolidGeometries{});
@@ -489,12 +489,12 @@ namespace osc
                 }
                 if (ui::draw_menu_item("Landmarks from CSV"))
                 {
-                    ActionLoadLandmarksFromCSV(m_State->updUndoable(), m_DocumentIdentifier);
+                    ActionLoadLandmarksFromCSV(m_State->getUndoableSharedPtr(), m_DocumentIdentifier);
                 }
                 if (m_DocumentIdentifier == TPSDocumentInputIdentifier::Source &&
                     ui::draw_menu_item("Non-Participating Landmarks from CSV"))
                 {
-                    ActionLoadNonParticipatingLandmarksFromCSV(m_State->updUndoable());
+                    ActionLoadNonParticipatingLandmarksFromCSV(m_State->getUndoableSharedPtr());
                 }
                 ui::end_popup();
             }